requirements.txt
pyverilog==1.2
networkx==3.0
gensim==4.3.1
torch==2.0.1
torch_geometric==2.3.1
ultralytics==8.0.0
klayout==0.28.6
pyyaml==6.0
pytest==7.2.0
pytest-mock==3.10.0
scikit-learn==1.2.2
Pillow==9.5.0
